fix: Modify file name to match naming convention
